backend/requirements.txt
flask
flask-cors
bcrypt
mysql-connector-python